Crea paquetes, reorganiza clases en ellos y crea interface TrabajoDao
