+++ app/src/main/java/gruporebechi_garcialozano_tibalt/dam/isi/frsf/lab03/OfertasListAdapter.java
@@ -14,6 +14,8 @@ import java.text.NumberFormat;
 import java.text.SimpleDateFormat;
 import java.util.List;
 
+import gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.model.Trabajo;
+
 /**
  * Created by diegogarcialozano on 19/09/17.
  */
