+++ app/src/main/java/gruporebechi_garcialozano_tibalt/dam/isi/frsf/lab03/dao/TrabajoDao.java
@@ -1,0 +1,30 @@
+package gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.dao;
+
+import java.util.List;
+
+import gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.model.Categoria;
+import gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.model.Trabajo;
+
+/**
+ * Created by diegogarcialozano on 24/10/17.
+ */
+
+public interface TrabajoDao {
+    /**
+     *  Lista de las categorias predefinidas
+     * @return List<Categoria>
+     */
+    List<Categoria> listaCategoria();
+
+    /**
+     * Crea un nuevo trabajo
+     * @param p Trabajo a crear
+     */
+    void crearOferta(Trabajo p);
+
+    /**
+     * Retorna la lista de trabajos
+     * @return List<Trabajo>
+     */
+    List<Trabajo> listaTrabajos();
+}
