+++ app/src/main/java/gruporebechi_garcialozano_tibalt/dam/isi/frsf/lab03/exceptions/BadInputException.java
@@ -1,4 +1,4 @@
-package gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03;
+package gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.exceptions;
 
 /**
  * Created by diegogarcialozano on 12/10/17.
@@ -8,7 +8,7 @@ public class BadInputException extends Exception {
 
     private String msg;
 
-    BadInputException(String msg) {
+    public BadInputException(String msg) {
         super(msg);
         this.msg = msg;
     }
