+++ app/src/main/java/gruporebechi_garcialozano_tibalt/dam/isi/frsf/lab03/model/Trabajo.java
@@ -1,10 +1,9 @@
-package gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03;
+package gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
 import android.util.Log;
 
-import java.io.Serializable;
 import java.util.Date;
 import java.util.Random;
