+++ app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.MainActivity"
+    tools:context="gruporebechi_garcialozano_tibalt.dam.isi.frsf.lab03.activities.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <ListView
